Move action builder types to internal/types
